# docs/api.yaml
openapi: 3.1.0
info:
  title: TalkItOut API
  version: 1.0.0
  description: AI-powered support system API for Singapore students
servers:
  - url: http://localhost:4000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /tasks:
    get:
      summary: Get user tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tasks retrieved
    post:
      summary: Create task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                subject:
                  type: string
                priority:
                  type: string
                  enum: [low, med, high]
                dueAt:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Task created
...